revert: 69a486a886b2b3892e2fa38e0f21226ad8917a37
we might accidentally use isEmpty on a pointer, which is not valid because the zero value of a pointer is nil, which reflect.DeepEqual will report as a false negative if we pass a &T{} even though this should be also be empty based on our desired logic. We want &T{} also to be marked as "empty" as nil.